+++ contracts/tokens/DummyTokenB.sol
@@ -6,12 +6,12 @@ contract DummyTokenB is StandardToken {
   string public name = "TokenB";
   string public symbol = "TB";
 
-  function DummyTokenB(uint256 _value) {
+  function DummyTokenB(uint _value) {
     balances[msg.sender] = _value;
     totalSupply = _value;
   }
 
-  function setBalance(uint256 _value) {
+  function setBalance(uint _value) {
     balances[msg.sender] = _value;
   }
 }
