+++ contracts/util/ExchangeMathUtil.sol
@@ -10,12 +10,12 @@ contract ExchangeMathUtil is SafeMath {
   /// @param filledValueM Value of order already filled.
   /// @return Amount of tokenM to fill.
   function getFillValueM(
-    uint256 valueM,
-    uint256 fillValueM,
-    uint256 filledValueM)
+    uint valueM,
+    uint fillValueM,
+    uint filledValueM)
     constant
     internal
-    returns (uint256 toFillValueM)
+    returns (uint toFillValueM)
   {
     if (safeAdd(filledValueM, fillValueM) > valueM) {
       fillValueM = safeSub(valueM, filledValueM);
@@ -28,10 +28,10 @@ contract ExchangeMathUtil is SafeMath {
   /// @param fillValueM Amount of tokenM to be filled.
   /// @param target Value to calculate partial.
   /// @return Partial value of target.
-  function getPartialValue(uint256 valueM, uint256 fillValueM, uint256 target)
+  function getPartialValue(uint valueM, uint fillValueM, uint target)
     constant
     internal
-    returns (uint256 partial)
+    returns (uint partial)
   {
     assert(fillValueM <= valueM);
     assert(!(target < 10**3 && target * fillValueM % valueM != 0)); // throw if rounding error > 0.1%
